Cleanup race conditions on banning logic
